+++ NewsServiceApi/NewsServiceApi/BL/Service/NewsService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using NewsServiceApi.DAL.Repositories;
-using NewsServiceApi.DAL.Model;
+using NewsServiceApi.DAL.Models;
 using NewsServiceApi.BL.DTO;
 
 namespace NewsServiceApi.BL.Service
@@ -24,7 +24,7 @@ namespace NewsServiceApi.BL.Service
         }
 
         //TODO: can I use lambda expression instead recall ?
-        public async Task<NewsDTO> GetByIdAsync(long id)
+        public async Task<NewsDTO> GetByIdAsync(int id)
         {
             News news =  await _newsRepository.GetByIdAsync(id);
             return AutoMapper.Mapper.Map<News, NewsDTO>(news);
