+++ NewsServiceApi/NewsServiceApi/DAL/Repositories/INewsRepository.cs
@@ -1,15 +1,16 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using NewsServiceApi.DAL.Model;
+using NewsServiceApi.DAL.Models;
 
 namespace NewsServiceApi.DAL.Repositories
 {
     public interface INewsRepository
     {
-        Task<News> GetByIdAsync(long Id);
+        Task<News> GetByIdAsync(int Id);
         Task<IEnumerable<News>> GetAllAsync();
+       // Task<IEnumerable<News>> GetByCategoryAsync();
         Task CreateAsync(News news);
         Task UpdateAsync(News news);
-        Task DeleteAsync(long Id);
+        Task DeleteAsync(int Id);
     }
 }
